[fix] do-activity State completion. Still broken.
In the current state if a do-activity finishes after
a State's children have reached a FinalState, the
completion event posted by the do-activity thread
doesn't wake up the StateMachine, thus the completion
transition on the State never fire... (unless
another event is posted). The fix is to have a unified
standard Event and Completion Event queue. This
is slated to be done on a different branch.